Moved code to configmap.go
